판매중인 물품조회 진행중

+++ mini/src/main/java/co/jin/pj/user/User.java
@@ -13,38 +13,43 @@ public class User {
 	static Scanner scn = new Scanner(System.in);
 	static UserService service = new UserServiceImpl();
 	static List<UserVO> list = new ArrayList<UserVO>();
-	
+
 	public static String loginUserId;
-	
+
 	public static void login() {
 		UserVO vo = new UserVO();
 		clearScreen();
-		System.out.println("======================== 로 그 인 ========================");
+		System.out.println("로 그 인 =============================================================================================================================================================");
 		System.out.print("ID >>");
 		String userId = scn.nextLine();
 		vo.setUserId(userId);
-		System.out.print("비밀번호 >>");
-		vo.setUserPassword(scn.nextLine());
+		if (service.selectUser2(vo).getBlackList().equals("Y")) {
+			System.out.println("블랙리스트 접속 차단");
+		} else {
+			System.out.print("비밀번호 >>");
+			vo.setUserPassword(scn.nextLine());
 
-		int result = service.loginUser(vo);
+			int result = service.loginUser(vo);
 
-		if (result == 1) {
-			loginUserId = userId;  
-			Menu.loginMenu();
+			if (result == 1) {
+				loginUserId = userId;
+				clearScreen();
+				Menu.loginMenu();
+			}
 		}
 	}
 
 	public static void userInsert() { /// 가입
 		UserVO vo = new UserVO();
 		clearScreen();
-		System.out.println("======================== 회 원 가 입 ========================");
+		System.out.println("회 원 가 입 =============================================================================================================================================================");
 		System.out.print("ID >>");
 		while (true) {
 			String userId = scn.nextLine();
 			vo.setUserId(userId);
 			if (!userId.equals(service.selectUser2(vo).getUserId())) {
 				break;
-			} else if(userId.equals(service.selectUser2(vo).getUserId())) {
+			} else if (userId.equals(service.selectUser2(vo).getUserId())) {
 				System.out.println("이미 가입된 ID입니다.");
 				System.out.print("다른 ID를 입력해주세요 >>");
 			}
@@ -77,21 +82,28 @@ public class User {
 	}
 
 	public static void userDelete() { /// 회원 탈퇴
-		System.out.println("======================== 회 원 탈 퇴 ========================");
+		System.out.println("회 원 탈 퇴 =============================================================================================================================================================");
 		UserVO vo = new UserVO();
-		System.out.print("삭제할 ID >>");
-		String userId = scn.nextLine();
+		String userId = User.loginUserId;
 		vo.setUserId(userId);
+
 		System.out.print("비밀번호 >>");
 		String userPassword = scn.nextLine();
-		vo.setUserPassword(userPassword);
 
-		service.deleteUser(vo);
+		if (userPassword.equals(service.selectUser2(vo).getUserPassword())) {
+			vo.setUserPassword(userPassword);
+			System.out.println(User.loginUserId + " 회원 탈퇴 되었습니다.");
+			service.deleteUser(vo);
+			Menu.main();
+		} else if (!userPassword.equals(service.selectUser2(vo).getUserPassword())) {
+			System.out.println("비밀번호가 틀렸습니다.");
+		}
+
 	}
 
 	public static void blackList() {
 		clearScreen();
-		System.out.println("======================== 블 랙 리 스 트 명 단 ========================");
+		System.out.println("블 랙 리 스 트 명 단 =============================================================================================================================================================");
 		list = service.blackListSelect();
 		for (UserVO vo : list) {
 			System.out.printf("ID : %s  닉네임 : %s \n", vo.getUserId(), vo.getUserNickname());
@@ -100,24 +112,23 @@ public class User {
 
 	public static void selectUser() {
 		clearScreen();
-		System.out.println("======================== 회 원 정 보 조 회 ========================");
+		System.out.println("회 원 정 보 조 회 =============================================================================================================================================================");
 		UserVO vo = new UserVO();
 		UserVO vo2 = new UserVO();
-		System.out.print("조회할 ID >>");
-		String id = scn.nextLine();
+		String id = User.loginUserId;
 		vo2.setUserId(id);
 		System.out.print("비밀번호 >>");
 		String password = scn.nextLine();
 		vo2.setUserPassword(password);
-		
-		
-		if(id.equals(service.selectUser(vo2).getUserId())) {
-		vo.setUserId(id);
-		vo.setUserPassword(password);
-		vo = service.selectUser(vo);
-		
-		System.out.printf("ID : %s  연락처 : %s  주소 : %s  닉네임 : %s \n", vo.getUserId(), vo.getUserPhone(),
-				vo.getUserAddress(), vo.getUserNickname());
+//		vo3 = service.selectUser(vo2);
+
+		if (id.equals(service.selectUser(vo2).getUserId())) {
+			vo.setUserId(id);
+			vo.setUserPassword(password);
+			vo = service.selectUser(vo);
+
+			System.out.printf("ID : %s  연락처 : %s  주소 : %s  닉네임 : %s \n", vo.getUserId(), vo.getUserPhone(),
+					vo.getUserAddress(), vo.getUserNickname());
 		} else {
 			System.out.println("ID, 비밀번호가 일치하지 않습니다.");
 		}
@@ -127,29 +138,35 @@ public class User {
 	public static void userUpdate() { /// 회원정보 수정
 		UserVO vo = new UserVO();
 		clearScreen();
-		System.out.println("======================== 회 원 정 보 수 정 ========================");
-		System.out.print("현재 ID >>");
-		String userId = scn.nextLine();
+		System.out.println("회 원 정 보 수 정 =============================================================================================================================================================");
+		System.out.println("현재 ID는 '" + User.loginUserId + "' 입니다.");
+		String userId = User.loginUserId;
 		vo.setUserId(userId);
 		System.out.println();
 		System.out.print("현재 비밀번호 >>");
 		String userPassword = scn.nextLine();
 		vo.setUserPassword(userPassword);
-		System.out.print("수정할 비밀번호 >>");
-		String userPassword2 = scn.nextLine();
-		vo.setUserPassword(userPassword2);
-		System.out.print("수정할 연락처 >>");
-		String phone = scn.nextLine();
-		vo.setUserPhone(phone);
-		System.out.print("수정할 주소 >>");
-		String address = scn.nextLine();
-		vo.setUserAddress(address);
-		System.out.print("수정할 닉네임 >>");
-		vo.setUserNickname(scn.nextLine());
-
-		service.updateUser(vo);
+		if (userPassword.equals(service.selectUser2(vo).getUserPassword())) {
+			System.out.print("수정할 비밀번호 >>");
+			String userPassword2 = scn.nextLine();
+			vo.setUserPassword2(userPassword2);
+			System.out.print("수정할 연락처 >>");
+			String phone = scn.nextLine();
+			vo.setUserPhone(phone);
+			System.out.print("수정할 주소 >>");
+			String address = scn.nextLine();
+			vo.setUserAddress(address);
+			System.out.print("수정할 닉네임 >>");
+			vo.setUserNickname(scn.nextLine());
+			clearScreen();
+			System.out.println("수정 완료");
+			service.updateUser(vo);
+		} else {
+			clearScreen();
+			System.out.println("비밀번호가 일치하지않습니다.");
+		}
 	}
-	
+
 	public static void clearScreen() {
 		for (int i = 0; i < 80; i++)
 			System.out.println("");
